Move EditorModel to the model file.
